pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bearing"
version = "0.1.0"
description = "Avennorth Bearing — CMDB health assessment platform for ServiceNow"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Proprietary"}
authors = [
    { name = "Avennorth", email = "engineering@avennorth.com" },
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "requests>=2.32.0",
    "anthropic>=0.40.0",
    "fpdf2>=2.8.0",
    "python-docx>=1.1.0",
    "matplotlib>=3.9.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "eval_type_backport>=0.3.0;python_version<'3.10'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "types-requests>=2.32.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
bearing = "bearing.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/bearing"]

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["bearing"]
omit = ["src/tests/*"]

[tool.coverage.report]
fail_under = 80
